+++ R/summary.R
@@ -1,0 +1,11 @@
+#' @include occumb.R
+NULL
+
+setGeneric("summary")
+
+setMethod("summary", signature(object = "occumbFit"),
+    function(object, ...) {
+        summary(object@fit, ...)
+    }
+)
+