+++ NEWS.md
@@ -1,6 +1,7 @@
 # occumb (development version)
 * Change occumbData class specification and occumbData() function.
 * Add occumb() function.
+* Add methods for occumbFit class: plot, print, summary.
 
 # occumb v0.1.0
 * Initial development of occumb package.